application: split schema
